ClimbingFilter: Refactoring
